+++ apps/builder/app/shared/nano-states/canvas.ts
@@ -1,8 +1,8 @@
 import { atom, computed } from "nanostores";
 import type { Instance, Instances } from "@webstudio-is/sdk";
+import type { FontWeight } from "@webstudio-is/fonts";
 import { $instances } from "./instances";
 import type { InstanceSelector } from "../tree-utils";
-import type { FontWeight } from "@webstudio-is/fonts";
 
 export type TextToolbarState = {
   selectionRect: undefined | DOMRect;
